Fix spellcheck false positive
The codespell spellchecker tool is used to detect occurrences of commonly misspelled words in the project files.

The project readme references the acronym "ALS" (ambient light sensor). This happens to be in the codespell misspelled
word dictionary as a misspelling of "also", and so causes a failure of the spellcheck.

The false positive is resolved by adding the word to the ignore list in the codespell configuration.